fix(clippy): clean -D warnings across workspace via xtask
CI uses `cargo run -p xtask -- clippy --workspace --all-targets -- -D warnings`,
which sets up vcpkg + MSVC env so ffmpeg-sys-next finds libavutil. Running
clippy this way (instead of bare `cargo clippy`) surfaced a real set of
lints that bare invocation hadn't been compiling far enough to see.

Pre-existing (not from earlier bughunt commits):
* gpu-mem/src/lib.rs:310 — `map_or(true, |b| …)` → `is_none_or(|b| …)`
  (clippy::unnecessary_map_or). Cleaner Rust 1.82+ ergonomics.
* pt-denoise-oidn/src/lib.rs:422 — `weights_dir.as_ref().map(|p| p.as_path())`
  → `as_deref()` (clippy::option_as_ref_deref).
* pt-denoise-oidn/src/lib.rs:743 — `bytes_per_row % 256 != 0` →
  `!bytes_per_row.is_multiple_of(256)` (clippy::manual_is_multiple_of).
* pt-denoise-oidn/src/lib.rs:657 — extracted `type HwC4Input = (Tensor, Buffer, u64)`
  to satisfy clippy::type_complexity. Doc-commented what each tuple slot is.
* pt-denoise-oidn/src/lib.rs:95-98 — rewrote doc bullet that started with
  `+ tile-plan compute) is paid…` to use commas, fixing
  clippy::doc_lazy_continuation that misparsed the `+` as a bullet.
* pt-megakernel/src/compute.rs:1472 — added blank line between bullet list
  and trailing paragraph (clippy::doc_lazy_continuation).
* src/app/treemap_view.rs:717 — annotated `commit_marquee_at` with
  `#[allow(clippy::too_many_arguments)]` + a doc note that the 8-arg
  signature is intentional dedup of two parallel call sites at 427/449
  (folding into a struct would push the per-arg thread-through into
  `select_objects_in_rect` without removing it).
* src/app/settings/denoiser.rs:153 — `.clamp_to_range(false)` was already
  deprecated in eframe 0.34 (so was `clamp_existing_to_range` that this
  replaced). Switched to `.clamping(egui::SliderClamping::Never)` per the
  deprecation hint.

From earlier bughunt commits (self-introduced regressions):
* pt-wavefront/src/wavefront/pipeline.rs:356 — dropped redundant
  `TILE_SLOT_STRIDE as u64` cast (const is already `u64`).
  This was my regression from the tile_offset overflow fix in
  commit 7e93470.

Verified: `cargo run -p xtask -- clippy --workspace --all-targets -- -D warnings`
exits 0 with zero diagnostics. Workspace tests pass via `cargo run -p xtask -- test --debug`
(no failures across the 10 test binaries — 0 unit tests defined).

+++ crates/pt-wavefront/src/wavefront/pipeline.rs
@@ -353,7 +353,7 @@ impl WavefrontPipeline {
     /// and the math goes through u64 so release builds never silently wrap.
     pub fn tile_offset(&self, idx: u32) -> u32 {
         debug_assert!(idx < self.tile_capacity, "tile_idx out of range");
-        let off = idx as u64 * TILE_SLOT_STRIDE as u64;
+        let off = idx as u64 * TILE_SLOT_STRIDE;
         debug_assert!(off <= u32::MAX as u64, "tile offset {off} overflows u32");
         off as u32
     }
